Extract citation logic from chat router into services/citations
